fix 'loop' bug

+++ src/main/scala/de/htwg/scalala/advancedDsl/MusicVar.scala
@@ -1,3 +1,3 @@
 package de.htwg.scalala.advancedDsl
 
-case class MusicVar(identifier: String, position: Int = 0)
+case class MusicVar(identifier: String, position: Long = 0)
